backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
websockets
python-multipart
httpx